chore: move methods to helper class
