+++ enderio/src/main/java/com/enderio/enderio/client/content/travel/TravelTargetRendering.java
@@ -87,44 +87,4 @@ public class TravelTargetRendering {
             event.getRenderState().setRenderData(DATA_KEY, renderStates);
         }
     }
-
-//    @SubscribeEvent
-//    public static void renderLevel(RenderLevelStageEvent.AfterWeather event) {
-//        List<ExtractTravelTarget> renderStates = event.getLevelRenderState().getRenderData(DATA_KEY);
-//        if (renderStates == null) {
-//            return;
-//        }
-//
-//        RenderTarget mainTarget = Minecraft.getInstance().getMainRenderTarget();
-//        if (mainTarget.getDepthTexture() != null) {
-//            RenderSystem.getDevice().createCommandEncoder().clearDepthTexture(mainTarget.getDepthTexture(), 1.0);
-//        }
-//
-//        renderStates.sort(Comparator.comparingDouble(ExtractTravelTarget::distanceSquared).reversed());
-//        var previousColorTextureOverride = RenderSystem.outputColorTextureOverride;
-//        var previousDepthTextureOverride = RenderSystem.outputDepthTextureOverride;
-//        try {
-//            RenderSystem.outputColorTextureOverride = mainTarget.getColorTextureView();
-//            RenderSystem.outputDepthTextureOverride = mainTarget.getDepthTextureView();
-//
-//            for (ExtractTravelTarget state : renderStates) {
-//                PoseStack poseStack = event.getPoseStack();
-//                poseStack.pushPose();
-//                Vec3 projectedView = event.getLevelRenderState().cameraRenderState.pos;
-//                poseStack.translate(
-//                    state.target.pos().getX() - projectedView.x,
-//                    state.target.pos().getY() - projectedView.y,
-//                    state.target.pos().getZ() - projectedView.z);
-//
-//                // needed for smooth rendering
-//                // the boolean value controls whether it's still smooth while the game world is
-//                // paused (e.g. /tick freeze)
-//                render(state.target, event.getLevelRenderer(), poseStack, state.distanceSquared, state.active, state.partialTick);
-//                poseStack.popPose();
-//            }
-//        } finally {
-//            RenderSystem.outputColorTextureOverride = previousColorTextureOverride;
-//            RenderSystem.outputDepthTextureOverride = previousDepthTextureOverride;
-//        }
-//    }
 }
